phase-5/backend/requirements.txt
# FastAPI and ASGI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
mangum>=0.17.0

# Database
sqlmodel>=0.0.22
psycopg2-binary>=2.9.9

# Environment
python-dotenv>=1.0.0

# Authentication (JWT)
PyJWT>=2.8.0
cryptography>=42.0.0

# OpenAI and AI Agent
openai>=1.0.0

# MCP (Model Context Protocol)
# Note: Install from PyPI when available, or use local package

# HTTP client for async operations
httpx>=0.27.0
